backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
mysql-connector-python==8.2.0
python-dotenv==1.0.0
cloudinary==1.36.0
replicate==0.22.0
requests==2.31.0
huggingface-hub==0.20.2
pydantic==2.5.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
aiohttp==3.9.1
